+++ CHANGELOG.md
@@ -14,6 +14,7 @@ with the exception that 0.x versions can break between minor versions.
 - Add support for word boundaries and zero-length fancy patterns inside variable lookbehinds (at top level) (#216)
 - Add support for `\R` to mean general newline, matching all common line break characters, treating `\r\n` atomically (#220)
 - Add support for subroutine calls (including recursion up to 20 levels deep, matching Oniguruma behavior) `\g<1>` (#230)
+- Add support for Oniguruma's absent repeater `(?~abc)` (#233)
 ### Changed
 - `RegexBuilder` can now build multiple patterns with the same options (#213)
 ### Fixed